+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.14.0 (in development)
 * Add Spanish and Japanese language support.
+* Add MySQL authentication support.
 
 ## 0.13.0 (2024-06-27)
 * Add feature to hide quotation marks in File Search.